# protobuf/swagger/protobuf/api/user.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "protobuf/api/user.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "mughieams.evermosassessment.v1.UserService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/users": {
      "get": {
        "operationId": "GetUsers",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/mughieams.evermosassessment.v1.GetUsersResponse"
            }
          }
        },
        "tags": [
          "mughieams.evermosassessment.v1.UserService"
        ]
      },
      "post": {
        "operationId": "Register",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/mughieams.evermosassessment.v1.MessageResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/mughieams.evermosassessment.v1.RegisterRequest"
            }
          }
        ],
        "tags": [
          "mughieams.evermosassessment.v1.UserService"
        ]
      }
    },
    "/v1/users/login": {
      "post": {
        "operationId": "Login",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/mughieams.evermosassessment.v1.LoginResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/mughieams.evermosassessment.v1.LoginRequest"
            }
          }
        ],
        "tags": [
          "mughieams.evermosassessment.v1.UserService"
        ]
      }
    },
    "/v1/users/{id}": {
      "patch": {
        "operationId": "UpdateUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/mughieams.evermosassessment.v1.MessageResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/mughieams.evermosassessment.v1.UserService.UpdateUserBody"
            }
          }
        ],
        "tags": [
          "mughieams.evermosassessment.v1.UserService"
        ]
      }
    }
  },
  "definitions": {
    "mughieams.evermosassessment.v1.GetUsersResponse": {
      "type": "object",
      "properties": {
        "users": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/mughieams.evermosassessment.v1.User"
          }
        }
      }
    },
    "mughieams.evermosassessment.v1.LoginRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "required": [
        "password"
      ]
    },
    "mughieams.evermosassessment.v1.LoginResponse": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        }
      }
    },
    "mughieams.evermosassessment.v1.MessageResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "readOnly": true
        }
      }
    },
    "mughieams.evermosassessment.v1.RegisterRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "mughieams.evermosassessment.v1.User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        }
      }
    },
    "mughieams.evermosassessment.v1.UserService.UpdateUserBody": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        }
      }
    }
  }
}